+++ src/utils/api-utils/api-constants.js
@@ -12,6 +12,7 @@ export const ENCODED_STRING_END_CHAR = global.encodeURIComponent(
   STRING_END_CHAR
 );
 export const ENCODED_QUOTE_CHAR = global.encodeURIComponent('"');
+export const ENCODED_EQUAL_CHAR = global.encodeURIComponent("=");
 
 export const OMNI_TEXT_SEARCH_TYPE = Symbol("omni-text");
 export const LIKE_TEXT_SEARCH_TYPE = Symbol("like-text");
@@ -23,5 +24,7 @@ export const DATE_SEARCH_TYPE = Symbol("date");
 export const DATETIME_SEARCH_TYPE = Symbol("datetime");
 export const FULL_ID_SEARCH_TYPE = Symbol("full-id");
 export const LIKE_ID_SEARCH_TYPE = Symbol("like-id");
+
+export const URI_QUERY_TYPE = Symbol("uri-query");
 
 export const DATE_SEARCH_TYPES = [DATE_SEARCH_TYPE, DATETIME_SEARCH_TYPE];